Added new library for parsing JSON files, wrapped name column, inverted sorts, started styling

+++ src/main/java/data/Item.java
@@ -53,7 +53,7 @@ public class Item {
     }
 
     public void setValue(String value) {
-        this.value = BigDecimal.valueOf(Long.parseLong(value));
+        this.value = BigDecimal.valueOf(Double.parseDouble(value));
     }
 
     public void setSerial(String serial) {
